+++ service/src/main/java/org/whispersystems/textsecuregcm/subscriptions/SubscriptionPrice.java
@@ -4,6 +4,6 @@
  */
 package org.whispersystems.textsecuregcm.subscriptions;
 
-import java.math.BigDecimal;
-
-public record SubscriptionPrice(String currency, BigDecimal amount) {}
+/// @param currency The upper-case 3-character currency specifier of the price
+/// @param amount   The amount in the currency's minor unit (as defined by stripe)
+public record SubscriptionPrice(String currency, long amount) {}
